Move global tools to material
